feat(uom): purchase-vs-base UoM separation + decimal quantities (SCO-143)
Adds the three-tier UoM model so operators receive in the truck-driver-
facing unit (BAG, PACK, CASE) while inventory + recipes stay in the
canonical base unit (LB, EA). Conversion is applied at receipt time so
all downstream consumers (preflight, reservations, picks, reports)
operate on one canonical quantity space.

Decision (Meatbag, 2026-05-28):
- Q1 lot granularity → aggregated lot per receipt (KISS; per-unit lots
  deferred until regulator asks)
- Q2 decimal qty   → YES, all qty columns become Float (granular to
  ~15 significant digits via float64; 0.001 precision documented for
  the user-visible contract; Numeric(12,3) on the Postgres cutover)
- Q3 UoM authority → SKU base UoM wins; recipe-line UoM gets converted
  at preflight via the SCO-51 v2 BOM conversion path

Backend — schema additions:
- SKU.purchase_uom (VARCHAR(10), default '') — optional packaging unit
- SKU.base_per_purchase_unit (FLOAT, default 1.0) — conversion factor
- Runtime additive migration via _ensure_columns() in main.py
- ASNLineOut now exposes purchase_uom + base_uom + base_per_purchase_unit
  so the receipt-editor UI can render conversion previews without a
  second fetch

Backend — type changes (Integer → Float across the quantity surface):
- Lot.quantity, LotGenealogy.quantity_consumed
- SKU.reorder_point, SKU.safety_stock
- ASNLine.expected_qty / received_qty
- ReceiptLine.qty_received / qty_variance
- OrderLine.qty_ordered / qty_picked, Pick.qty_picked
- WorkOrder.target_qty, WorkOrderReservation.qty_reserved
- All matching Pydantic schemas (LotOut, SKUDetailOut, InventoryKPIs,
  AdjustRequest, BelowSafetyRow, ASNLineOut, ReceiptLineIn, ReceiptOut,
  PutawaySuggestion, ASNLineIn, OrderLineOut, PickAssignmentRequest,
  PickOut, ConsolidationLotPlan, ConsolidationPlan, PackingSlipLine,
  OrderLineIn, WorkOrderCreate, ReservationOut, WorkOrderOut,
  WorkOrderCompleteRequest, SKURow, SKUCreate)
- SQLite's type-affinity model makes the column-type change a no-op at
  the storage layer. On Postgres cutover (future), explicit ALTER
  COLUMN TYPE migrations needed — _ensure_columns() only handles ADD
  COLUMN. Flagged inline in main.py.

Backend — receipt service:
- services/receiving.py create_receipt multiplies qty_received by
  sku.base_per_purchase_unit at lot creation so lot.quantity is the
  base-unit derivative.
- putaway_suggestions also applies the conversion factor so the
  suggested slot capacities are checked against the stocked qty in
  base UoM (matches lot.quantity).

Frontend:
- scripts/creators.js createSKU form gains "Purchase unit (optional)" +
  "Base units per purchase unit" fields. Defaults preserve existing
  behavior (blank purchase_uom + factor 1.0 = no conversion).
- ASN modal SKU options now show purchase context next to the SKU code
  ("GARLIC-001 · Garlic powder · qty in BAG of 50 LB") + body copy
  explains that qty is in purchase UoM and the server converts at
  receipt time.
- UoMs are uppercased client-side before submit for consistency.

Tests:
- tests/test_purchase_uom_conversion.py — 6 cases:
  - SKU defaults (no packaging, factor 1.0)
  - SKU with packaging persists + round-trips through GET /skus
  - Receipt converts: 10 BAG × 50 LB/BAG = 500 LB stocked
  - Default SKUs (no packaging) behave exactly as pre-143
  - Decimal preservation: 0.5 BAG → 25.0 LB
  - UoM case normalization (currently passthrough — flagged)

Verification:
- ruff clean across the entire wms/ tree + new test file
- pytest 264/264 passing (was 258; +6 new cases, zero regressions)
- node -c clean on creators.js

Future work (queued as SCO-144 in the next session):
Mass-unit measurement-system conversion (LB↔KG↔OZ↔G) at the input
layer. Operator types "1000 KG" → frontend converts to 2204.62 LB →
backend stores canonical LB. Design + scope discussed in chat;
distinct from SCO-143 because it operates at the measurement-system
boundary, not the packaging boundary.

+++ backend/wms/models/production.py
@@ -18,7 +18,7 @@ MVP scope notes (each TODO is intentionally left here for future iteration):
 
 from datetime import UTC, datetime
 
-from sqlalchemy import DateTime, ForeignKey, Integer, String
+from sqlalchemy import DateTime, Float, ForeignKey, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from wms.db.base import Base
@@ -61,7 +61,8 @@ class WorkOrder(Base):
     # Snapshot of the recipe version at WO creation. When versioning lands,
     # this field is what protects a running WO from a mid-flight recipe edit.
     recipe_version_snapshot: Mapped[int] = mapped_column(Integer, default=1, nullable=False)
-    target_qty: Mapped[int] = mapped_column(Integer, nullable=False)
+    # SCO-143: target/actual yields in base UoM, decimal-capable.
+    target_qty: Mapped[float] = mapped_column(Float, nullable=False)
     # State machine: draft → reserved → running → completed.
     # `cancel` is the only exit from reserved/running short of completed.
     status: Mapped[str] = mapped_column(String(20), default="draft", index=True, nullable=False)
@@ -78,4 +79,5 @@ class WorkOrderReservation(Base):
     id: Mapped[int] = mapped_column(primary_key=True)
     work_order_id: Mapped[int] = mapped_column(ForeignKey("work_orders.id"), index=True, nullable=False)
     lot_id: Mapped[int] = mapped_column(ForeignKey("lots.id"), index=True, nullable=False)
-    qty_reserved: Mapped[int] = mapped_column(Integer, nullable=False)
+    # SCO-143: reservation qty in base UoM, decimal-capable (e.g., 1.5 lb).
+    qty_reserved: Mapped[float] = mapped_column(Float, nullable=False)
